requirements.txt
requests
click
Pillow
